Refactoring: Addressing various pylint/isort/flake8/tox errors
