+++ tox.ini
@@ -25,6 +25,7 @@ deps=
 	coveralls
 usedevelop=true
 commands_pre=
+	python -m pip install --require-hashes -r requirements.txt -r test-requirements.txt
 commands=
 	pytest --cov-report=html --cov=exodus_gw {posargs}
